store monitroing data in chunked file on server, statistic improvemnet

+++ projects/copper-monitoring/copper-monitoring-core/src/main/java/de/scoopgmbh/copper/monitoring/core/statistic/CountAggregateFunction.java
@@ -3,7 +3,7 @@ package de.scoopgmbh.copper.monitoring.core.statistic;
 import java.util.List;
 
 
-public class CountAggregateFunction<T> implements AggregateFunction<T,TimeValuePair<Long>>{
+public class CountAggregateFunction<T> implements AggregateFunction<T,TimeValuePair<Double>>{
 	private static final long serialVersionUID = -714208820060517375L;
 
 	public CountAggregateFunction() {
@@ -11,8 +11,8 @@ public class CountAggregateFunction<T> implements AggregateFunction<T,TimeValueP
 	}
 
 	@Override
-	public TimeValuePair<Long> doAggregate(List<T> value, TimeframeGroup<T,TimeValuePair<Long>> usedGroup) {
-		return new TimeValuePair<Long>(usedGroup.to,(long) value.size());
+	public TimeValuePair<Double> doAggregate(List<T> value, TimeframeGroup<T,TimeValuePair<Double>> usedGroup) {
+		return new TimeValuePair<Double>(usedGroup.to,(double) value.size());
 	}
 
 }
